refactor: 다이얼로그 컴포넌트로 분리
